unify the way webgl and webgl2 get patched
