+++ packages/rrweb/src/record/observers/canvas/webgl.ts
@@ -3,13 +3,112 @@ import {
   blockClass,
   CanvasContext,
   canvasMutationCallback,
+  canvasMutationParam,
   IWindow,
   listenerHandler,
   Mirror,
-  SerializedWebGlArg,
 } from '../../../types';
 import { hookSetter, isBlocked, patch } from '../../../utils';
 import { saveWebGLVar, serializeArgs } from './serialize-args';
+
+const pendingCanvasMutations = new Map<
+  HTMLCanvasElement,
+  canvasMutationParam[]
+>();
+
+// FIXME: total hack here, we need to find a better way to do this
+function flushPendingCanvasMutations(
+  cb: canvasMutationCallback,
+  mirror: Mirror,
+) {
+  pendingCanvasMutations.forEach(
+    (values: canvasMutationParam[], canvas: HTMLCanvasElement) => {
+      const id = mirror.getId((canvas as unknown) as INode);
+      flushPendingCanvasMutationFor(canvas, id, cb);
+    },
+  );
+  requestAnimationFrame(() => flushPendingCanvasMutations(cb, mirror));
+}
+
+function flushPendingCanvasMutationFor(
+  canvas: HTMLCanvasElement,
+  id: number,
+  cb: canvasMutationCallback,
+) {
+  const values = pendingCanvasMutations.get(canvas);
+  if (!values || id === -1) return;
+
+  values.forEach((p) => cb({ ...p, id }));
+  pendingCanvasMutations.delete(canvas);
+}
+
+function patchGLPrototype(
+  prototype: WebGLRenderingContext | WebGL2RenderingContext,
+  type: CanvasContext,
+  cb: canvasMutationCallback,
+  blockClass: blockClass,
+  mirror: Mirror,
+): listenerHandler[] {
+  const handlers: listenerHandler[] = [];
+
+  const props = Object.getOwnPropertyNames(prototype);
+
+  for (const prop of props) {
+    try {
+      if (typeof prototype[prop as keyof typeof prototype] !== 'function') {
+        continue;
+      }
+      const restoreHandler = patch(prototype, prop, function (original) {
+        return function (this: typeof prototype, ...args: Array<unknown>) {
+          const result = original.apply(this, args);
+          saveWebGLVar(result);
+          if (!isBlocked((this.canvas as unknown) as INode, blockClass)) {
+            const id = mirror.getId((this.canvas as unknown) as INode);
+
+            const recordArgs = serializeArgs([...args]);
+            const mutation = {
+              id,
+              type,
+              property: prop,
+              args: recordArgs,
+            };
+
+            if (id === -1) {
+              if (!pendingCanvasMutations.has(this.canvas))
+                pendingCanvasMutations.set(this.canvas, []);
+
+              pendingCanvasMutations
+                .get(this.canvas as HTMLCanvasElement)!
+                .push(mutation);
+            } else {
+              // flush all pending mutations
+              flushPendingCanvasMutationFor(this.canvas, id, cb);
+              cb(mutation);
+            }
+          }
+
+          return result;
+        };
+      });
+      handlers.push(restoreHandler);
+    } catch {
+      const hookHandler = hookSetter<typeof prototype>(prototype, prop, {
+        set(v) {
+          cb({
+            id: mirror.getId((this.canvas as unknown) as INode),
+            type,
+            property: prop,
+            args: [v],
+            setter: true,
+          });
+        },
+      });
+      handlers.push(hookHandler);
+    }
+  }
+
+  return handlers;
+}
 
 export default function initCanvasWebGLMutationObserver(
   cb: canvasMutationCallback,
@@ -18,62 +117,32 @@ export default function initCanvasWebGLMutationObserver(
   mirror: Mirror,
 ): listenerHandler {
   const handlers: listenerHandler[] = [];
-  const props = Object.getOwnPropertyNames(win.WebGLRenderingContext.prototype);
-  for (const prop of props) {
-    try {
-      if (
-        typeof win.WebGLRenderingContext.prototype[
-          prop as keyof WebGLRenderingContext
-        ] !== 'function'
-      ) {
-        continue;
-      }
-      const restoreHandler = patch(
-        win.WebGLRenderingContext.prototype,
-        prop,
-        function (original) {
-          return function (
-            this: WebGLRenderingContext,
-            ...args: Array<unknown>
-          ) {
-            if (!isBlocked((this.canvas as unknown) as INode, blockClass)) {
-              setTimeout(() => {
-                const recordArgs = serializeArgs([...args]);
-                cb({
-                  id: mirror.getId((this.canvas as unknown) as INode),
-                  type: CanvasContext.WebGL,
-                  property: prop,
-                  args: recordArgs,
-                });
-              }, 0);
-            }
 
-            const result = original.apply(this, args);
-            saveWebGLVar(result);
-            return result;
-          };
-        },
-      );
-      handlers.push(restoreHandler);
-    } catch {
-      const hookHandler = hookSetter<WebGLRenderingContext>(
-        win.WebGLRenderingContext.prototype,
-        prop,
-        {
-          set(v) {
-            cb({
-              id: mirror.getId((this.canvas as unknown) as INode),
-              type: CanvasContext.WebGL,
-              property: prop,
-              args: [v],
-              setter: true,
-            });
-          },
-        },
-      );
-      handlers.push(hookHandler);
-    }
+  // TODO: replace me
+  requestAnimationFrame(() => flushPendingCanvasMutations(cb, mirror));
+
+  handlers.push(
+    ...patchGLPrototype(
+      win.WebGLRenderingContext.prototype,
+      CanvasContext.WebGL,
+      cb,
+      blockClass,
+      mirror,
+    ),
+  );
+
+  if (typeof win.WebGL2RenderingContext !== 'undefined') {
+    handlers.push(
+      ...patchGLPrototype(
+        win.WebGL2RenderingContext.prototype,
+        CanvasContext.WebGL2,
+        cb,
+        blockClass,
+        mirror,
+      ),
+    );
   }
+
   return () => {
     handlers.forEach((h) => h());
   };
